fix: 좋아요 수(likeCount) int → long으로 변경

+++ src/main/java/com/sprint/mission/monew/domain/comment/dto/response/CommentLikeResponse.java
@@ -12,7 +12,7 @@ public record CommentLikeResponse(
     UUID commentUserId,
     String commentUserNickname,
     String commentContent,
-    int commentLikeCount,
+    long commentLikeCount,
     Instant commentCreatedAt
 ) {
